Added new submission streaming function

+++ mira/post.go
@@ -6,10 +6,10 @@ func (p *PostListing) GetChildren() []PostListingChild {
 	return p.Data.Children
 }
 
-func (plc *PostListingChild) GetAge() int64 {
-	return time.Now().UTC().Unix() - plc.Data.CreatedUtc
+func (plc *PostListingChild) GetAge() float64 {
+	return float64(time.Now().UTC().Unix()) - plc.Data.CreatedUtc
 }
 
-func (plc *PostListingChild) GetTimeCreated() int64 {
+func (plc *PostListingChild) GetTimeCreated() float64 {
 	return plc.Data.CreatedUtc
 }
